+++ tests/acceptance/fixtures/order_by/alpha_fixture.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+function test_alpha_one() {
+  assert_same 1 1
+}
+
+function test_alpha_two() {
+  assert_same 2 2
+}